int is 32 bit on Windows (cause trouble for hybrid_IA)

+++ python/damask/_grid.py
@@ -754,8 +754,8 @@ class Grid:
         # materials: 1
 
         """
-        offset_ = np.array(offset,int) if offset is not None else np.zeros(3,int)
-        cells_ = np.array(cells,int) if cells is not None else self.cells
+        offset_ = np.array(offset,np.int64) if offset is not None else np.zeros(3,np.int64)
+        cells_ = np.array(cells,np.int64) if cells is not None else self.cells
 
         canvas = np.full(cells_,np.nanmax(self.material) + 1 if fill is None else fill,self.material.dtype)
 
@@ -797,7 +797,7 @@ class Grid:
 
         >>> import numpy as np
         >>> import damask
-        >>> g = damask.Grid(np.zeros([32]*3,int), np.ones(3)*1e-4)
+        >>> g = damask.Grid(np.zeros([32]*3,int),np.ones(3)*1e-4)
         >>> g.mirror('xy',True)
         cells : 64 x 64 x 32
         size  : 0.0002 x 0.0002 x 0.0001 m³
